bs_backend/requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dotenv==0.9.9
google-auth==2.38.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
urllib3==2.3.0
